Crate problem019

Problem 19 (Counting Sundays)

Solution detail

Given that 1 Jan 1900 was a Monday, we can work out which day the following months began on by keeping track of the current day, and advancing it according to how many days were in the month which just passed, modulo 7. All this requires is being able to work out how many days were in a given month in a given year.

Then simply do this over a given range of months and count how many times the day at the start of the month was Sunday.

Functions

answer
Solve the problem, returning the answer as a String

days_in_month
Returns how many days were in the given month in the given year, where 0 -> January, 1 -> February and so on.

main

solve
Find the number of months from 1901 to the given year which began on the given day, where 0 -> Monday, 1 -> Tuesday and so on.